requirements.txt
label-studio
pdf2image
pdfminer.six
